+++ src/generators/helpMessage.js
@@ -1,10 +1,12 @@
 function createHelpCommandsList(botInstance) {
     const commands = [];
     for(let i = 0; i < botInstance.commandsList.length; i++) {
-        const commandObj = { inline: false };
-        commandObj.title = botInstance.config.prefix + botInstance.commandsList[i].name;
-        commandObj.value = botInstance.stringProcessor(botInstance.commandsList[i].description);
-        commands.push(commandObj);
+        if(!botInstance.commandsList[i].hidden) {
+            const commandObj = { inline: false };
+            commandObj.title = botInstance.config.prefix + botInstance.commandsList[i].name;
+            commandObj.value = botInstance.stringProcessor(botInstance.commandsList[i].description);
+            commands.push(commandObj);
+        }
     }
     return commands;
 }